Changed the getters and setters of data to a struct based approach
With the old approach it was easy for data to get desynchronized since not every packet contains the same information
In this new approach, each sentence is written to its own struct which has a getter. The getter writes a copy
of the data to a pointer to its respective struct. The data is expected to be the latest that has been received from the gps hardware.
There is currently no way to lock the data to prevent race conditions. Still exploring how best to deal with that